@@ -1,6 +1,6 @@
 {
   "name": "Skyclerk",
-  "version": "3.0.0",
+  "version": "3.0.3",
   "author": "Cloudmanic Labs, LLC",
   "homepage": "https://skyclerk.com",
   "scripts": {
@@ -13,61 +13,51 @@
   },
   "private": true,
   "dependencies": {
-    "@agm/core": "^1.0.0",
-    "@angular/common": "^7.2.2",
-    "@angular/core": "^7.2.2",
-    "@angular/forms": "^7.2.2",
-    "@angular/http": "^7.2.2",
-    "@angular/platform-browser": "^7.2.2",
-    "@angular/platform-browser-dynamic": "^7.2.2",
-    "@angular/router": "^7.2.2",
-    "@capacitor/android": "^1.3.0",
-    "@capacitor/cli": "^1.3.0",
-    "@capacitor/core": "^1.3.0",
-    "@capacitor/ios": "^1.3.0",
-    "@ionic-native/core": "^5.5.0",
-    "@ionic-native/file-transfer": "^5.5.0",
-    "@ionic-native/in-app-purchase-2": "^5.25.0",
-    "@ionic/angular": "^4.1.0",
-    "@types/mixpanel": "^2.14.2",
-    "ajv-keywords": "^3.4.0",
-    "cc.fovea.cordova.purchase": "^10.1.1",
-    "cordova-plugin-file": "^6.0.1",
-    "cordova-plugin-file-transfer": "^1.7.1",
+    "@agm/core": "^1.1.0",
+    "@angular/common": "~8.2.14",
+    "@angular/core": "~8.2.14",
+    "@angular/forms": "~8.2.14",
+    "@angular/platform-browser": "~8.2.14",
+    "@angular/platform-browser-dynamic": "~8.2.14",
+    "@angular/router": "~8.2.14",
+    "@capacitor/core": "2.0.1",
+    "@capacitor/ios": "^2.1.0",
+    "@ionic-native/core": "^5.26.0",
+    "@ionic-native/file-transfer": "^5.26.0",
+    "@ionic-native/in-app-purchase-2": "^5.26.0",
+    "@ionic-native/splash-screen": "^5.0.0",
+    "@ionic-native/status-bar": "^5.0.0",
+    "@ionic/angular": "^5.0.0",
     "core-js": "^2.5.4",
-    "help": "3.0.2",
-    "ios-deploy": "^1.9.4",
-    "ios-sim": "^8.0.1",
-    "mixpanel-browser": "^2.30.1",
-    "moment": "^2.24.0",
-    "rxjs": "~6.3.3",
-    "zone.js": "~0.8.29"
+    "mixpanel-browser": "^2.36.0",
+    "moment": "^2.25.3",
+    "rxjs": "~6.5.1",
+    "tslib": "^1.9.0",
+    "zone.js": "~0.9.1"
   },
   "devDependencies": {
-    "@angular-devkit/architect": "~0.12.3",
-    "@angular-devkit/build-angular": "~0.13.0",
-    "@angular-devkit/core": "~7.2.3",
-    "@angular-devkit/schematics": "~7.2.3",
-    "@angular/cli": "~7.3.1",
-    "@angular/compiler": "~7.2.2",
-    "@angular/compiler-cli": "~7.2.2",
-    "@angular/language-service": "~7.2.2",
-    "@ionic/angular-toolkit": "~1.4.0",
-    "@types/jasmine": "~2.8.8",
+    "@angular-devkit/build-angular": "~0.803.20",
+    "@angular/cli": "~8.3.23",
+    "@angular/compiler": "~8.2.14",
+    "@angular/compiler-cli": "~8.2.14",
+    "@angular/language-service": "~8.2.14",
+    "@capacitor/cli": "2.0.1",
+    "@ionic/angular-toolkit": "^2.1.1",
+    "@types/jasmine": "~3.3.8",
     "@types/jasminewd2": "~2.0.3",
-    "@types/node": "~10.14.2",
-    "codelyzer": "~4.5.0",
-    "jasmine-core": "~2.99.1",
+    "@types/node": "~8.9.4",
+    "codelyzer": "^5.0.0",
+    "jasmine-core": "~3.4.0",
     "jasmine-spec-reporter": "~4.2.1",
-    "karma": "~3.1.4",
+    "karma": "~4.1.0",
     "karma-chrome-launcher": "~2.2.0",
     "karma-coverage-istanbul-reporter": "~2.0.1",
-    "karma-jasmine": "~1.1.2",
-    "karma-jasmine-html-reporter": "^0.2.2",
+    "karma-jasmine": "~2.0.1",
+    "karma-jasmine-html-reporter": "^1.4.0",
     "protractor": "~5.4.0",
-    "ts-node": "~8.0.0",
-    "tslint": "~5.12.0",
-    "typescript": "~3.1.6"
+    "ts-node": "~7.0.0",
+    "tslint": "~5.15.0",
+    "typescript": "~3.4.3"
   },
   "description": "An Ionic project",
   "cordova": {
