@@ -1,6 +1,6 @@
 export const environment = {
 	production: true,
-	version: "3.0.2",
+	version: "3.0.3",
 	client_id: "XL8TeRGBdsUvvM3",
 	app_server: "https://app.skyclerk.com",
 	mixpanel_key: "5974bde1c90a5606add2695a17b2db10"
