@@ -4,10 +4,10 @@
 
 export const environment = {
 	production: false,
-	version: "3.0.2",
+	version: "3.0.3",
 	client_id: "abc123wootsmobile",
-	app_server: "http://127.0.0.1:9090",
-	//app_server: "http://192.168.1.100:9090",
+	//app_server: "http://127.0.0.1:9090",
+	app_server: "http://192.168.1.100:9090",
 	mixpanel_key: "",
 	//mixpanel_key: "5974bde1c90a5606add2695a17b2db10",
 };
